Added integrety checks to database functions.

+++ models/entry.ts
@@ -3,16 +3,19 @@ import { Entry } from "../types/types.ts";
 
 export function insertEntry(entry: Entry) {
   const db = new DatabaseSync("db/jotbot.db");
+  if (
+    !(db.prepare("PRAGMA integrity_check;").get()?.integrity_check === "ok")
+  ) throw new Error("JotBot Error: Databaes integrety check failed!");
   db.exec("PRAGMA foreign_keys = ON;");
   db.prepare(
-    `INSERT INTO entry_db (userId, timestamp, situation, automaticThoughts, emotionName, emotionEmoji, emotionDescription, selfiePath) VALUES (?, ?, ?, ?, ?, ?, ?, ?)`,
+    `INSERT INTO entry_db (userId, timestamp, situation, automaticThoughts, emotionName, emotionEmoji, emotionDescription, selfiePath) VALUES (?, ?, ?, ?, ?, ?, ?, ?);`,
   ).run(
     entry.userId,
     entry.timestamp,
     entry.situation,
     entry.automaticThoughts,
     entry.emotion.emotionName,
-    entry.emotion.emotionEmoji,
+    entry.emotion.emotionEmoji || null,
     entry.emotion.emotionDescription,
     entry.selfiePath,
   );
@@ -23,8 +26,11 @@ export function getEntriesByUserId(userId: number): Entry[] {
   const entries = [];
   try {
     const db = new DatabaseSync("db/jotbot.db");
+    if (
+      !(db.prepare("PRAGMA integrity_check;").get()?.integrity_check === "ok")
+    ) throw new Error("JotBot Error: Databaes integrety check failed!");
     const queryResults = db.prepare(
-      `SELECT * FROM entry_db WHERE userId = '${userId}' ORDER BY timestamp DESC`,
+      `SELECT * FROM entry_db WHERE userId = '${userId}' ORDER BY timestamp DESC;`,
     ).all();
     for (const e in queryResults) {
       const entry: Entry = {
@@ -55,7 +61,10 @@ export function getEntriesByUserId(userId: number): Entry[] {
 export function deleteEntry(entryId: number) {
   try {
     const db = new DatabaseSync("db/jotbot.db");
-    db.prepare(`DELETE FROM entry_db WHERE id = '${entryId}'`).run();
+    if (
+      !(db.prepare("PRAGMA integrity_check;").get()?.integrity_check === "ok")
+    ) throw new Error("JotBot Error: Databaes integrety check failed!");
+    db.prepare(`DELETE FROM entry_db WHERE id = '${entryId}';`).run();
     db.close();
   } catch (err) {
     console.log(`Failed to delete entry ${entryId} from entry_db: ${err}`);
